requirements.txt
asgiref~=3.6.0
django-autoslug==1.9.8
click==8.0.4
colorama==0.4.4
django-crispy-forms==1.14.0
django-debug-toolbar==3.4.0
django-ineo @ git+https://github.com/proft/django-ineo.git@fe421f01c74ff738407ee569d4db0a051ca04542
click==8.0.4
colorama==0.4.4
Django==3.2.7
django-autoslug==1.9.8
django-crispy-forms==1.14.0
django-debug-toolbar==3.4.0
django-ineo==0.5
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
nltk==3.7
Pillow==9.1.1
pytz~=2021.1
regex==2022.3.2
sqlparse~=0.4.1
tqdm==4.63.0
tzdata==2022.1
mysqlclient==2.1.1
pytils==0.4.1
django-ckeditor==6.5.1
numpy==1.23.3
Pillow==9.1.1
psycopg2-binary==2.9.4
pytz==2022.1
regex==2022.3.2
scipy==1.9.1
sqlparse==0.4.2
tqdm==4.63.0
tzdata==2022.1

pip~=22.3.1
wheel~=0.37.1
setuptools~=60.2.0
django-modeltranslation~=0.18.7
certifi~=2020.12.5
chardet~=4.0.0
idna~=2.10
urllib3~=1.26.3
requests~=2.25.1
six~=1.15.0
djangorestframework~=3.14.0